started songcrawler class refactoring
